requirements.txt
openai>=1.0.0
pillow>=12.0.0
qrcode>=7.4.0
